+++ automationCode/Changelog.md
@@ -5,11 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## 12 October 2022
+- Made changes to output layout of the bison file
+- Made changes to flex file regex to support new way of creating the grammar rules
+- Some changes to extract Param regex methods
 
 ## 10 October 2022
 - Improved findTokenValue method
 - Improved retrieving params from writeStr
 - Need to look into sprintf and how to make it work
+- Added new file for later. Want to refactor by moving related code to its own files
 
 ## 15 August 2022
 - Started From scratch with grammar auto generation